Helpers Update
TODO: перезапись фото с измененным exif

+++ Helpers/ImageProcessingService.cs
@@ -1,55 +1,38 @@
 ﻿using ExifLib;
 using PhotoPreparation.ViewModels;
 using Serilog;
-using System;
-using System.Collections.Generic;
 using System.Drawing.Imaging;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PhotoPreparation.Helpers
 {
     public class ImageProcessingService(SettingsViewModel settingsViewModel)
     {
         private static readonly object lockObject = new();
+
         public event Action<string>? StatusTextChanged;
         public event Action<double>? ProgressValueChanged;
 
         public async Task<bool> ProcessImagesAsync(string inputFolderPath, int maxWidth, int maxHeight)
         {
             var selectedFontSize = GetFontSize(settingsViewModel.SelectedFontSizeIndex);
-
             var tempOutputFolderPath = Path.Combine(inputFolderPath, "ProcessedTemp");
-            var finalOutputFolderPath = inputFolderPath;
+            var finalOutputFolderPath = settingsViewModel.DeleteOriginalPhotos ? tempOutputFolderPath : Path.Combine(inputFolderPath, MessageConstants.OutputFolderName);
 
-            if (settingsViewModel.DeleteOriginalPhotos)
-                finalOutputFolderPath = tempOutputFolderPath;
-            else
+            if (!settingsViewModel.DeleteOriginalPhotos && Directory.Exists(finalOutputFolderPath) && Directory.EnumerateFileSystemEntries(finalOutputFolderPath).Any())
             {
-                finalOutputFolderPath = Path.Combine(inputFolderPath, MessageConstants.OutputFolderName);
-
-                if (Directory.Exists(finalOutputFolderPath) && Directory.EnumerateFileSystemEntries(finalOutputFolderPath).Any())
+                var result = MessageBox.Show("Папка Обработанные фото уже существует и не пустая. Перезаписать обработанные файлы?", "Предупреждение", MessageBoxButtons.YesNo);
+                if (result == DialogResult.No)
                 {
-                    var result = MessageBox.Show("Папка Обработанные фото уже существует и не пустая. Перезаписать обработанные файлы?", "Предупреждение", MessageBoxButtons.YesNo);
-                    if (result == DialogResult.No)
-                    {
-                        OnStatusTextChanged(MessageConstants.CancelledByUser);
-                        return false;
-                    }
-                    Directory.Delete(finalOutputFolderPath, true);
+                    OnStatusTextChanged(MessageConstants.CancelledByUser);
+                    return false;
                 }
+                Directory.Delete(finalOutputFolderPath, true);
             }
-
-
 
             string[] allowedExtensions = [".jpg", ".jpeg", ".png"];
 
             Directory.CreateDirectory(finalOutputFolderPath);
-
-            // Получение списка уже обработанных файлов в выходной папке
-            HashSet<string> processedFiles = new(Directory.GetFiles(finalOutputFolderPath));
 
             string[] files = Directory.GetFiles(inputFolderPath, "*.*", SearchOption.TopDirectoryOnly)
                                     .Where(filePath => allowedExtensions.Contains(Path.GetExtension(filePath)?.ToLower()))
@@ -57,7 +40,8 @@ namespace PhotoPreparation.Helpers
 
             var counterSuccess = 0;
             var counterFailure = 0;
-            StringBuilder stringBuilder = new();
+
+            var font = new Font("Arial", selectedFontSize, FontStyle.Bold);
 
             await Task.Run(() =>
             {
@@ -74,12 +58,10 @@ namespace PhotoPreparation.Helpers
                                 var newWidth = Math.Min(originalImage.Width, maxWidth);
                                 var newHeight = Math.Min(originalImage.Height, maxHeight);
 
-                                using Image resizedImage = ImageProcessingService.ResizeImage(originalImage, newWidth, newHeight);
+                                using Image resizedImage = ResizeImage(originalImage, newWidth, newHeight);
                                 using Graphics g = Graphics.FromImage(resizedImage);
-                                var font = new Font("Arial", selectedFontSize, FontStyle.Bold);
 
                                 var watermark = dateTime.ToString("dd/MM/yyyy HH:mm");
-
                                 SizeF textSize = g.MeasureString(watermark, font);
 
                                 // Координаты для отрисовки текста в правом нижнем углу
@@ -99,9 +81,6 @@ namespace PhotoPreparation.Helpers
                                     resizedImage.Save(Path.Combine(finalOutputFolderPath, fileName), ImageFormat.Jpeg);
                                     resizedImage.Dispose();
                                 }
-
-                                lock (lockObject)
-                                    processedFiles.Add(filePath);
                             }
                         }
 
@@ -109,26 +88,25 @@ namespace PhotoPreparation.Helpers
                         {
                             lock (lockObject)
                             {
-                                var filePathWithoutExtension = Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {counterFailure + 1}{MessageConstants.JpgExtension}");
+                                var filePathWithoutExtension = Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {counterFailure + 1} {filePath}{MessageConstants.JpgExtension}");
+
                                 if (!File.Exists(filePathWithoutExtension))
-                                    File.Copy(filePath, Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}"));
-                                processedFiles.Add(Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}"));
+                                    File.Copy(filePath, Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure} {filePath}{MessageConstants.JpgExtension}"));
                             }
                         }
-
                     }
                     catch (ExifLibException ex)
                     {
                         Log.Error(ex, $"Ошибка работы с EXIF: {filePath} {ex.Message}");
 
-
                         lock (lockObject)
                         {
-                            var filePathWithoutExtension = Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {counterFailure + 1}{MessageConstants.JpgExtension}");
+                            var filePathWithoutExtension = Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {counterFailure + 1} {filePath}{MessageConstants.JpgExtension}");
+
                             if (!File.Exists(filePathWithoutExtension))
                                 File.Copy(filePath, Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}"));
-                            processedFiles.Add(filePath);
-                            OnStatusTextChanged($"Ошибка при работе с EXIF файла {Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}")}");
+
+                            OnStatusTextChanged($"Ошибка при работе с EXIF файла {Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure} {filePath}{MessageConstants.JpgExtension}")}");
                         }
                     }
 
@@ -136,14 +114,7 @@ namespace PhotoPreparation.Helpers
                     {
                         Log.Error(ex, $"Неизвестная ошибка: {filePath} {ex.Message}");
 
-                        lock (lockObject)
-                        {
-                            var filePathWithoutExtension = Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {counterFailure + 1}{MessageConstants.JpgExtension}");
-                            if (!File.Exists(filePathWithoutExtension))
-                                File.Copy(filePath, Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}"));
-                            processedFiles.Add(filePath);
-                            OnStatusTextChanged($"Неизвестная ошибка при работе с файлом {Path.Combine(finalOutputFolderPath, $"{MessageConstants.NoMetaData} {++counterFailure}{MessageConstants.JpgExtension}")}");
-                        }
+                        MessageBox.Show($"Произошла неизвестная ошибка при обработке {filePath}. Файл не обработан.\nСкинь логи в Dropbox (в моих документах, папка Photo Preparation)");
                     }
 
                     finally
@@ -153,9 +124,10 @@ namespace PhotoPreparation.Helpers
                     }
                 });
             });
-            return true;
 
+            return true;
         }
+
         public static Bitmap ResizeImage(Image image, int maxWidth, int maxHeight)
         {
             var ratio = Math.Min((double)maxWidth / image.Width, (double)maxHeight / image.Height);
@@ -163,6 +135,7 @@ namespace PhotoPreparation.Helpers
             var newHeight = (int)(image.Height * ratio);
 
             var newImage = new Bitmap(newWidth, newHeight);
+
             using var g = Graphics.FromImage(newImage);
             g.DrawImage(image, 0, 0, newWidth, newHeight);
 
@@ -195,13 +168,8 @@ namespace PhotoPreparation.Helpers
             };
         }
 
-        public void OnStatusTextChanged(string text)
-        {
-            StatusTextChanged?.Invoke(text);
-        }
-        public void OnProgressValueChanged(double value)
-        {
-            ProgressValueChanged?.Invoke(value);
-        }
+        public void OnStatusTextChanged(string text) => StatusTextChanged?.Invoke(text);
+
+        public void OnProgressValueChanged(double value) => ProgressValueChanged?.Invoke(value);
     }
 }
